--- Assets/Scirpts/UI/Function/GmaeOver.cs
@@ -1,20 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class GmaeOver : MonoBehaviour
-{
-    private Canvas canvas;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        canvas = GetComponent<Canvas>();
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
