--- Assets/Scirpts/UI/Function/GmaeOver.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8eee0dff06aefb140b19d45a2afe2a04
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
